Move app_manager.public_webform_submissions -> public_webforms.submissions
